shared/requirements.txt
# Shared dependencies for all services
fastapi>=0.104.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
redis>=5.0.0
python-multipart>=0.0.6
aiohttp>=3.9.0


# Security
bleach==6.1.0
python-magic==0.4.27

# Monitoring
redis==5.0.1
